requirements.txt
torch~=1.12.1
numpy~=1.23.5
scipy~=1.10.0
soundfile~=0.11.0
librosa~=0.9.2
h5py~=3.8.0
PyYAML~=6.0
thop~=0.1.1.post2209072238
tqdm~=4.65.0
pystoi~=0.3.3
pypesq~=1.2.4